finished package refactor
