+++ command-side/src/main/java/com/pankesh/productcommand/Application.java
@@ -1,4 +1,4 @@
-package com.soagrowers.productcommand;
+package com.pankesh.productcommand;
 
 import java.util.List;
