+++ command-side/src/main/java/com/pankesh/productcommand/commands/MarkProductAsSaleableCommand.java
@@ -1,4 +1,4 @@
-package com.soagrowers.productcommand.commands;
+package com.pankesh.productcommand.commands;
 
 import org.axonframework.commandhandling.TargetAggregateIdentifier;
