+++ common-things/src/main/java/com/pankesh/productevents/events/ProductSaleableEvent.java
@@ -1,4 +1,4 @@
-package com.soagrowers.productevents.events;
+package com.pankesh.productevents.events;
 
 
 public class ProductSaleableEvent extends AbstractEvent {
