+++ common-things/src/main/java/com/pankesh/utils/Asserts.java
@@ -1,4 +1,4 @@
-package com.soagrowers.utils;
+package com.pankesh.utils;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
