+++ common-things/src/test/java/com/pankesh/utils/AssertsTest.java
@@ -1,4 +1,4 @@
-package com.soagrowers.utils;
+package com.pankesh.utils;
 
 import org.junit.Before;
 import org.junit.Test;
