+++ config-service/src/main/java/com/pankesh/ConfigServiceApplication.java
@@ -1,4 +1,4 @@
-package com.soagrowers;
+package com.pankesh;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
