+++ discovery-service/src/main/java/com/pankesh/DiscoveryServiceApplication.java
@@ -1,4 +1,4 @@
-package com.soagrowers;
+package com.pankesh;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
