+++ gateway-service/src/main/java/com/pankesh/prefilters/SimpleLoggingPreFilter.java
@@ -1,4 +1,4 @@
-package com.soagrowers.prefilters;
+package com.pankesh.prefilters;
 
 import com.netflix.zuul.ZuulFilter;
 import com.netflix.zuul.context.RequestContext;
