+++ integration-test/src/integration-test/java/com/pankesh/productintegrationtests/EndToEndProductTest.java
@@ -1,8 +1,8 @@
-package com.soagrowers.productintegrationtests;
+package com.pankesh.productintegrationtests;
 
 
 
-import com.soagrowers.utils.Statics;
+import com.pankesh.utils.Statics;
 import org.apache.http.HttpStatus;
 import org.hamcrest.Matchers;
 import org.junit.After;
@@ -17,7 +17,7 @@ import java.util.UUID;
 import java.util.concurrent.TimeUnit;
 
 import static com.jayway.restassured.RestAssured.given;
-import static com.soagrowers.utils.Statics.*;
+import static com.pankesh.utils.Statics.*;
 
 /**
  * Created by ben on 24/02/16.
