+++ integration-test/src/integration-test/java/com/pankesh/utils/Statics.java
@@ -1,4 +1,4 @@
-package com.soagrowers.utils;
+package com.pankesh.utils;
 
 /**
  * Created by ben on 24/02/16.
