+++ query-side/src/main/java/com/pankesh/productquery/handlers/EventLoggingHandler.java
@@ -1,4 +1,4 @@
-package com.soagrowers.productquery.handlers;
+package com.pankesh.productquery.handlers;
 
 import org.axonframework.eventhandling.EventHandler;
 import org.slf4j.Logger;
@@ -6,9 +6,9 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
-import com.soagrowers.productevents.events.ProductAddedEvent;
-import com.soagrowers.productevents.events.ProductSaleableEvent;
-import com.soagrowers.productevents.events.ProductUnsaleableEvent;
+import com.pankesh.productevents.events.ProductAddedEvent;
+import com.pankesh.productevents.events.ProductSaleableEvent;
+import com.pankesh.productevents.events.ProductUnsaleableEvent;
 
 /**
  * Handler's (a.k.a. Listeners) can be used to react to events and perform associated
